Tweaking georgexsh's line number option.
I like the idea! But I didn't like the default. :)

I'd prefer to default this to clearing the line numbers, as that's more
in line with the "distraction free" roots of the plugin. I'd originally
planned to drop the `no` from the option name to reverse it's meaning
for consistency with the rest of the options. I changed my mind while
writing it, and turned the option into an action:
`g:vimroom_clear_line_numbers`. This seems easier to understand, and
more in line with how I think about the feature.

Closes Issue #14.